\name{fglm-package}
\alias{fglm-package}
\alias{fglm}
\docType{package}
\title{
\packageTitle{fglm}
}
\description{
\packageDescription{fglm}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fglm}
\packageIndices{fglm}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{fglm}

Maintainer: \packageMaintainer{fglm}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
