% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ee.R
\name{fit_ee}
\alias{fit_ee}
\title{Fit model with latent exponential generalized linear model with log-link}
\usage{
fit_ee(
  y,
  yupp,
  X,
  lam = 1e-05,
  alpha = 0,
  pen_factor = c(0, rep(1, ncol(X) - 1)),
  maxit = rep(100, 3),
  tol = rep(1e-08, 2),
  method = "prox_newt",
  b = rep(0, ncol(X)),
  L = 10,
  verbose = FALSE,
  acc = TRUE
)
}
\arguments{
\item{y}{A vector of observed responses (see details).}

\item{yupp}{A vector of upper endpoints of the interval corresponding to y
(see details)}

\item{X}{A matrix of predictors whose i:th row corresponds to the i:th
element in y.}

\item{lam}{A vector of penalty parameters to fit the model for}

\item{alpha}{A scalar weight for elastic net (1 = lasso, 0 = ridge)}

\item{pen_factor}{A vector of coefficient-specific penalty weights; defaults
to 0 for fitst element of b and 1 for the remaining.}

\item{maxit}{A vector of maximum number of iterations (see details)}

\item{tol}{A vector of tolerances for FISTA or proximal Newton termination
(see details)}

\item{method}{Set to "fista" (FISTA) or "prox_newt" (proximal Newton)}

\item{b}{A vector of initial values for regression coefficients}

\item{L}{A scalar which if method = "fista" sets step-size 1 / L, which
guarantees convergence if the objective function has an L-Lipschitz gradient}

\item{verbose}{A logical which if TRUE means additional information may be
printed}

\item{acc}{A logical indicating whether to use acceleration when the FISTA
algorithm is used}
}
\value{
A matrix where each row correspond to a value of lam and the columns
are coefficient estimates (1 - p), the value of lam (p + 1), the number of
iterations required (p + 2), and whether zero is in the sub-differential at
the final iterate (p + 3)
}
\description{
{
  Minimizes an elastic net-penalized negative log-likelihood for finitely
  supported data from a latent exponential generalized linear model
  whose natural parameter is exp(X  b), implying its mean is exp(-X
  b).
}
}
\details{
{
  The model assumes latent responses are generated from an exponential
  generalized linear model with mean exp(-X b). The data are intervals in
  which the latent responses fell, [y, yupp), and the predictors.

  If method = "fista", then only the first elements of maxit and tol are
  used. If method = "prox_newt", then the first element of maxit is the
  maximum number of Newton iterations, the second is the maximum number of
  line search iterations for each Newton update, and the third is the maximum
  number of coordinate descent iterations within each Newton update. The
  first elemen of tol is for terminating the Newton iterations and the second
  for terminating the coordinate descent updates within each Newton
  iteration.


}
}
