% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.r
\name{generate_norm}
\alias{generate_norm}
\title{Generate data}
\usage{
generate_norm(X, b, d = 1, ymax = 5, ymin = -5, sigma = 1)
}
\arguments{
\item{X}{An n x p design matrix}

\item{b}{A vector of p regression coefficients}

\item{d}{A scalar controlling the coarseness of the support, where d = 1
means integer support (see details)}

\item{ymax}{An upper bound on the observable response (see details)}

\item{ymin}{A lower bound on the observable response}

\item{sigma}{Standard deviation of latent normal variable.}
}
\value{
A matrix with n rows and 3 columns; the first is the lower endpoint
 of the observed interval and the second the upper endpoint (see details).
 The third is the unobservable response.
}
\description{
{
Interval censored responses from the normal distribution with mean X %*% b.
}
}
\details{
{
  An unobservable response is normally distributed with mean X %*% b and 
  variance 1. Its support (-Inf, Inf) is partitioned in intervals
  (-Inf, ymin), [ymin, -k d), ... [-d, 0), [0, d), ... [ld, ymax),
  [ymax, Inf), where k = floor(-ymin / d) and l = floor(ymax / d).
  The returned interval [y, yupp) is that including the unobservable
  response.
 }
}
