% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.r
\name{hessian_icnet}
\alias{hessian_icnet}
\title{Evaluate the Hessian of log-likelihood}
\usage{
hessian_icnet(
  Y,
  X,
  b = NULL,
  s = NULL,
  theta = NULL,
  fix_s = TRUE,
  distr = "norm"
)
}
\arguments{
\item{Y}{Matrix of intervals containing latent responses.}

\item{X}{Model matrix.}

\item{b}{Vector (p x 1) of regression coefficients.}

\item{s}{Latent error standard deviation.}

\item{theta}{Parameter vector equal to [1/s, b'/s].}

\item{fix_s}{If TRUE, s is assumed known}

\item{distr}{Distribution of latent responses; "ee" for exponential.
distribution with log-link or "norm" for normal distribution with identity
link.}
}
\value{
A matrix of second order derivatives of the log-likelihood, with
respect to theta or [s, b']' depending on which arguments are supplied
}
\description{
Evaluate the Hessian of log-likelihood
}
