% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icnet_flex.R
\name{icnet_flex}
\alias{icnet_flex}
\title{Elastic-net regression for general interval censored latent variables}
\usage{
icnet_flex(
  M,
  Z,
  theta,
  lam = 1e-05,
  alpha = 0,
  pen_factor = NULL,
  box_constr = NULL,
  L = 10,
  maxit = rep(100, 3),
  tol = rep(1e-08, 2),
  method = "prox_newt",
  distr = "norm",
  verbose = FALSE,
  acc = TRUE,
  nfold = 1
)
}
\arguments{
\item{M}{Matrix (n x 2) of interval endpoints.}

\item{Z}{Model matrix (2n x d).}

\item{theta}{Starting values for parameter vector (d x 1).}

\item{lam}{Vector (d x 1) of penalty parameters.}

\item{alpha}{Scalar weight for elastic net (1 = lasso, 0 = ridge).}

\item{pen_factor}{Vector (d x 1) of coefficient-specific penalty weights.}

\item{box_constr}{matrix (d x 2) of box constraints. Can be -Inf (first col.)
or Inf (second col.).}

\item{L}{Scalar setting step-size 1 / L if method = "fista".}

\item{maxit}{Vector of maximum number of iterations.}

\item{tol}{Vector of tolerances for terminating algorithm.}

\item{method}{Method to use; "fista" or "prox_newt" (proximal Newton).}

\item{distr}{Distribution of latent responses; "ee" for exponential.
distribution with log-link or "norm" for normal distribution with identity
link.}

\item{verbose}{Logical indicating whether additional information should be
printed during fitting.}

\item{acc}{Logical indicating whether to use acceleration if
method = "fista".}

\item{nfold}{Number of folds in k-fold cross-validation; 1 corresponds
to no cross-validation.}
}
\value{
A matrix where each row correspond to a value of lam and the columns
are coefficient estimates (1 to d), the value of lam (d + 1), the number of
iterations required (d + 2), and whether zero is in the sub-differential at
the final iterate (d + 3)
}
\description{
{
  Minimizes an elastic net-penalized negative log-likelihood for interval
  censored latent variables using accelerated proximal gradient descent or 
  proximal Newton.
}
}
\details{
{
 Fit a model more general than icnet, but with fewer options, in particular,
 there is no cross-validation option.
 
 The likelihood for the ith observation is, for a log-concave cdf R,
  
    log(R(bi) - R(ai))
    
 where ai = M[ii, 1] + Z[ii, ] %*% theta and bi = M[ii, 2] + Z[ii + 1, ] %*%
 theta.
 
 If method = "fista", then only the first elements of maxit and tol are
 used. If method = "prox_newt", then the first element of maxit is the
 maximum number of Newton iterations, the second is the maximum number of
 line search iterations for each Newton update, and the third is the maximum
 number of coordinate descent iterations within each Newton update. The
 first element of tol is for terminating the Newton iterations and the second
 for terminating the coordinate descent updates within each Newton
 iteration.}
}
