% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.r
\name{obj_fun_icnet}
\alias{obj_fun_icnet}
\title{Evaluate objective function and its derivatives}
\usage{
obj_fun_icnet(
  Y,
  X,
  lam = 1e-05,
  alpha = 0,
  pen_factor = NULL,
  b,
  s = NULL,
  distr = "norm",
  order = 0
)
}
\arguments{
\item{Y}{Matrix of intervals containing latent responses.}

\item{X}{Model matrix.}

\item{lam}{Vector (d x 1) of penalty parameters.}

\item{alpha}{Scalar weight for elastic net (1 = lasso, 0 = ridge).}

\item{pen_factor}{Vector (d x 1) of coefficient-specific penalty weights.}

\item{b}{Vector (p x 1) of regression coefficients.}

\item{s}{Latent error standard deviation.}

\item{distr}{Distribution of latent responses; "ee" for exponential.
distribution with log-link or "norm" for normal distribution with identity
link.}

\item{order}{Order of derivatives to compute (0, 1, or 2)}
}
\value{
A list with elements "obj", "grad", and "hessian".
}
\description{
Evaluate objective function and its derivatives
}
\details{
{
 When order = 0, the gradient and Hessian elements of the return list are set
 to all zeros, and similarly for the Hessian when order = 1.
 
 The derivatives are with respect to theta = [1 / s, b' / s]'

 The sub-gradient returned is that obtained by taking the sub-gradient of the
 absolute value to equal zero at zero. When no element of b is zero, this is
 the usual gradient. The Hessian returned is that of the smooth part of the
 objective function; that is, the average negative log-likelihood plus the L2
 penalty only. When no element of b is zero, this is the Hessian of the
 objective function
}
}
