% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.r
\name{generate_ee}
\alias{generate_ee}
\title{Generate discretized and truncated data from a latent exponential
generalized linear model with natural parameter exp(X b)}
\usage{
generate_ee(X, b, d = 1, ymax = 10)
}
\arguments{
\item{X}{An n x p design matrix}

\item{b}{A vector of p regression coefficients}

\item{d}{A scalar controlling the coarseness of the support, where d = 1
means integer support (see details)}

\item{ymax}{An upper bound on the observable response (see details)}
}
\value{
A matrix with n rows and 2 columns; the first is the lower endpoint
 of the observed interval and the second the upper endpoint (see details)
}
\description{
Generate discretized and truncated data from a latent exponential
generalized linear model with natural parameter exp(X b)
}
\details{
{
  Data are generated by (i) draw y ~ Exp(exp(X b)), (ii) replace y by floor(y
  / d) * d, and (iii) replace every element of y greater than ymax by ymax;
  this gives the lower endpoint of the interval. Observe that the operation
  floor(y / d) * d sets y to the nearest smaller multiple of d.

  The upper endpoint of the observed interval is yupp = y + d, unless y =
  y_max in which case yupp = Inf.
 }
}
