% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icnet.R
\name{icnet}
\alias{icnet}
\title{Elastic-net penalized regression for interval censored responses}
\usage{
icnet(
  Y,
  X,
  lam = 1e-05,
  alpha = 0,
  pen_factor = NULL,
  b = NULL,
  s = NULL,
  fix_var = TRUE,
  box_constr = NULL,
  L = 10,
  maxit = rep(100, 3),
  tol = rep(1e-08, 2),
  method = "prox_newt",
  distr = "norm",
  verbose = FALSE,
  acc = TRUE,
  nfold = 1
)
}
\arguments{
\item{Y}{Matrix of intervals containing latent responses.}

\item{X}{Model matrix.}

\item{lam}{Vector (d x 1) of penalty parameters.}

\item{alpha}{Scalar weight for elastic net (1 = lasso, 0 = ridge).}

\item{pen_factor}{Vector (d x 1) of coefficient-specific penalty weights.}

\item{b}{Vector of initial values for regression coefficients.}

\item{s}{Scalar initial value for latent error standard deviation.}

\item{fix_var}{Logical indicating whether to treat latent error variance
as fixed and known.}

\item{box_constr}{matrix (d x 2) of box constraints. Can be -Inf (first col.)
or Inf (second col.).}

\item{L}{Scalar setting step-size 1 / L if method = "fista".}

\item{maxit}{Vector of maximum number of iterations.}

\item{tol}{Vector of tolerances for terminating algorithm.}

\item{method}{Method to use; "fista" or "prox_newt" (proximal Newton).}

\item{distr}{Distribution of latent responses; "ee" for exponential.
distribution with log-link or "norm" for normal distribution with identity
link.}

\item{verbose}{Logical indicating whether additional information should be
printed during fitting.}

\item{acc}{Logical indicating whether to use acceleration if
method = "fista".}

\item{nfold}{Number of folds in k-fold cross-validation; 1 corresponds
to no cross-validation.}
}
\value{
A matrix where each row correspond to a value of lam and the columns
are coefficient estimates (1 to d), the value of lam (d + 1), the number of
iterations required (d + 2), and whether zero is in the sub-differential at
the final iterate (d + 3)
}
\description{
{
  Minimizes an elastic net-penalized negative log-likelihood for interval
  censored responses using accelerated proximal gradient descent or 
  proximal Newton.
}
}
\details{
{
  
 The likelihood for the ith observation is, for a log-concave cdf R,
  
    log(R(bi) - R(ai))
    
 where ai = (Y[ii, 1] - t(X[ii, ]) %*% b) / s and 
 bi = (Y[ii, 2] - t(X[ii, ]) %*% b) / s. This is the log-probability of the
 event Y[ii, 1] < t(X[ii, ]) %*% b + s W[ii] < Y[ii, 2] when the elements
 of W are independent with cdf R.

 
 If method = "fista", then only the first elements of maxit and tol are
 used. If method = "prox_newt", then the first element of maxit is the
 maximum number of Newton iterations, the second is the maximum number of
 line search iterations for each Newton update, and the third is the maximum
 number of coordinate descent iterations within each Newton update. The
 first element of tol is for terminating the Newton iterations and the second
 for terminating the coordinate descent updates within each Newton
 iteration.}
}
