% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icnet.R
\name{icnet}
\alias{icnet}
\title{Elastic-net penalized regression for interval censored responses}
\usage{
icnet(
  Y,
  X,
  b = rep(0, ncol(X)),
  lam = 1e-05,
  alpha = 0,
  pen_factor = c(0, rep(1, ncol(X) - 1)),
  L = 10,
  maxit = rep(100, 3),
  tol = rep(1e-08, 2),
  method = "fista",
  distr = "norm",
  verbose = FALSE,
  acc = TRUE,
  nfold = 1
)
}
\arguments{
\item{Y}{A matrix (n x 2) with two columns corresponding to lower and
upper endpoints of response intervals (see details).}

\item{X}{A matrix (n x p) of predictors whose i:th row corresponds to the i:th
element in y.}

\item{b}{A vector of initial values for regression coefficients}

\item{lam}{A vector of penalty parameters to fit the model for}

\item{alpha}{A scalar weight for elastic net (1 = lasso, 0 = ridge)}

\item{pen_factor}{A vector of coefficient-specific penalty weights; defaults
to 0 for first element of b and 1 for the remaining.}

\item{L}{A scalar which if method = "fista" sets step-size 1 / L, which
guarantees convergence if the objective function has an L-Lipschitz gradient}

\item{maxit}{A vector of maximum number of iterations (see details)}

\item{tol}{A vector of tolerances for FISTA or proximal Newton termination
(see details)}

\item{method}{Set to "fista" (FISTA) or "prox_newt" (proximal Newton)}

\item{distr}{Set to "ee" for exponential distribution with log-link or "norm" for normal distribution.}

\item{verbose}{A logical which if TRUE means additional information may be
printed}

\item{acc}{A logical indicating whether to use acceleration when the FISTA
algorithm is used}

\item{nfold}{The number of folds in k-fold cross-validation; 1 corresponds
to no cross-validation.}
}
\value{
A matrix where each row correspond to a value of lam and the columns
are coefficient estimates (1 - p), the value of lam (p + 1), the number of
iterations required (p + 2), and whether zero is in the sub-differential at
the final iterate (p + 3)
}
\description{
{
  Minimizes an elastic net-penalized negative log-likelihood for interval
  censored responses using accelerated proximal gradient descent or 
  proximal Newton.
}
}
\details{
{
  The model assumes the observed data comprise predictor and an interval
  containing a latent, unobservable variable. The first column in Y is
  the lower endpoint of the interval and the second columns is the upper.

  If method = "fista", then only the first elements of maxit and tol are
  used. If method = "prox_newt", then the first element of maxit is the
  maximum number of Newton iterations, the second is the maximum number of
  line search iterations for each Newton update, and the third is the maximum
  number of coordinate descent iterations within each Newton update. The
  first element of tol is for terminating the Newton iterations and the second
  for terminating the coordinate descent updates within each Newton
  iteration.}
}
